@@ -1,6 +1,6 @@
 {
   "name": "react-dom",
-  "version": "15.4.2",
+  "version": "15.5.0-rc.2",
   "description": "React package for working with the DOM.",
   "main": "index.js",
   "repository": "facebook/react",
@@ -13,12 +13,13 @@
   },
   "homepage": "https://facebook.github.io/react/",
   "dependencies": {
-    "fbjs": "^0.8.1",
+    "fbjs": "^0.8.9",
     "loose-envify": "^1.1.0",
-    "object-assign": "^4.1.0"
+    "object-assign": "^4.1.0",
+    "prop-types": "~15.5.0"
   },
   "peerDependencies": {
-    "react": "^15.4.2"
+    "react": "^15.5.0-rc.2"
   },
   "files": [
     "LICENSE",
@@ -27,7 +28,8 @@
     "dist/",
     "lib/",
     "index.js",
-    "server.js"
+    "server.js",
+    "test-utils.js"
   ],
   "browserify": {
     "transform": [
