@@ -1,6 +1,6 @@
 {
   "name": "react-native-renderer",
-  "version": "15.4.2",
+  "version": "15.5.0-rc.2",
   "description": "React package for use inside react-native.",
   "main": "index.js",
   "repository": "facebook/react",
@@ -14,11 +14,11 @@
   },
   "homepage": "https://facebook.github.io/react-native/",
   "dependencies": {
-    "fbjs": "^0.8.1",
+    "fbjs": "^0.8.9",
     "object-assign": "^4.1.0"
   },
   "peerDependencies": {
-    "react": "^15.4.2"
+    "react": "^15.5.0-rc.2"
   },
   "files": [
     "LICENSE",
