@@ -1,6 +1,6 @@
 {
   "name": "react-test-renderer",
-  "version": "15.4.2",
+  "version": "15.5.0-rc.2",
   "description": "React package for snapshot testing.",
   "main": "index.js",
   "repository": "facebook/react",
@@ -15,17 +15,18 @@
   },
   "homepage": "https://facebook.github.io/react/",
   "dependencies": {
-    "fbjs": "^0.8.4",
+    "fbjs": "^0.8.9",
     "object-assign": "^4.1.0"
   },
   "peerDependencies": {
-    "react": "^15.4.2"
+    "react": "^15.5.0-rc.2"
   },
   "files": [
     "LICENSE",
     "PATENTS",
     "README.md",
     "index.js",
+    "shallow.js",
     "lib/"
   ]
 }
