@@ -12,6 +12,8 @@
 'use strict';
 
 var React = require('React');
+var propTypesFactory = require('prop-types/factory');
+var PropTypes = propTypesFactory(React.isValidElement);
 
 var ReactTransitionGroup = require('ReactTransitionGroup');
 var ReactCSSTransitionGroupChild = require('ReactCSSTransitionGroupChild');
@@ -52,9 +54,9 @@ class ReactCSSTransitionGroup extends React.Component {
   static propTypes = {
     transitionName: ReactCSSTransitionGroupChild.propTypes.name,
 
-    transitionAppear: React.PropTypes.bool,
-    transitionEnter: React.PropTypes.bool,
-    transitionLeave: React.PropTypes.bool,
+    transitionAppear: PropTypes.bool,
+    transitionEnter: PropTypes.bool,
+    transitionLeave: PropTypes.bool,
     transitionAppearTimeout: createTransitionTimeoutPropValidator('Appear'),
     transitionEnterTimeout: createTransitionTimeoutPropValidator('Enter'),
     transitionLeaveTimeout: createTransitionTimeoutPropValidator('Leave'),
