@@ -32,11 +32,11 @@ describe('ReactElement', () => {
     ReactTestUtils = require('ReactTestUtils');
     // NOTE: We're explicitly not using JSX here. This is intended to test
     // classic JS without JSX.
-    ComponentClass = React.createClass({
-      render: function() {
+    ComponentClass = class extends React.Component {
+      render() {
         return React.createElement('div');
-      },
-    });
+      }
+    };
   });
 
   afterEach(() => {
@@ -57,37 +57,7 @@ describe('ReactElement', () => {
     expect(element.props).toEqual({});
   });
 
-  it('should warn when `key` is being accessed on createClass element', () => {
-    spyOn(console, 'error');
-    var container = document.createElement('div');
-    var Child = React.createClass({
-      render: function() {
-        return <div> {this.props.key} </div>;
-      },
-    });
-    var Parent = React.createClass({
-      render: function() {
-        return (
-          <div>
-            <Child key="0" />
-            <Child key="1" />
-            <Child key="2" />
-          </div>
-        );
-      },
-    });
-    expect(console.error.calls.count()).toBe(0);
-    ReactDOM.render(<Parent />, container);
-    expect(console.error.calls.count()).toBe(1);
-    expect(console.error.calls.argsFor(0)[0]).toContain(
-      'Child: `key` is not a prop. Trying to access it will result ' +
-      'in `undefined` being returned. If you need to access the same ' +
-      'value within the child component, you should pass it as a different ' +
-      'prop. (https://fb.me/react-special-props)'
-    );
-  });
-
-  it('should warn when `key` is being accessed on ES class element', () => {
+  it('should warn when `key` is being accessed on composite element', () => {
     spyOn(console, 'error');
     var container = document.createElement('div');
     class Child extends React.Component {
@@ -95,8 +65,8 @@ describe('ReactElement', () => {
         return <div> {this.props.key} </div>;
       }
     }
-    var Parent = React.createClass({
-      render: function() {
+    class Parent extends React.Component {
+      render() {
         return (
           <div>
             <Child key="0" />
@@ -104,8 +74,8 @@ describe('ReactElement', () => {
             <Child key="2" />
           </div>
         );
-      },
-    });
+      }
+    }
     expect(console.error.calls.count()).toBe(0);
     ReactDOM.render(<Parent />, container);
     expect(console.error.calls.count()).toBe(1);
@@ -134,20 +104,20 @@ describe('ReactElement', () => {
   it('should warn when `ref` is being accessed', () => {
     spyOn(console, 'error');
     var container = document.createElement('div');
-    var Child = React.createClass({
-      render: function() {
+    class Child extends React.Component {
+      render() {
         return <div> {this.props.ref} </div>;
-      },
-    });
-    var Parent = React.createClass({
-      render: function() {
+      }
+    }
+    class Parent extends React.Component {
+      render() {
         return (
           <div>
             <Child ref="childElement" />
           </div>
         );
-      },
-    });
+      }
+    }
     expect(console.error.calls.count()).toBe(0);
     ReactDOM.render(<Parent />, container);
     expect(console.error.calls.count()).toBe(1);
@@ -255,12 +225,12 @@ describe('ReactElement', () => {
     var Component = React.createFactory(ComponentClass);
     var element;
 
-    var Wrapper = React.createClass({
-      render: function() {
+    class Wrapper extends React.Component {
+      render() {
         element = Component();
         return element;
-      },
-    });
+      }
+    }
 
     var instance = ReactTestUtils.renderIntoDocument(
       React.createElement(Wrapper)
@@ -312,19 +282,12 @@ describe('ReactElement', () => {
   it('allows static methods to be called using the type property', () => {
     spyOn(console, 'error');
 
-    var StaticMethodComponentClass = React.createClass({
-      statics: {
-        someStaticMethod: function() {
-          return 'someReturnValue';
-        },
-      },
-      getInitialState: function() {
-        return {valueToReturn: 'hi'};
-      },
-      render: function() {
+    class StaticMethodComponentClass extends React.Component {
+      render() {
         return React.createElement('div');
-      },
-    });
+      }
+    }
+    StaticMethodComponentClass.someStaticMethod = () => 'someReturnValue';
 
     var element = React.createElement(StaticMethodComponentClass);
     expect(element.type.someStaticMethod()).toBe('someReturnValue');
@@ -334,11 +297,11 @@ describe('ReactElement', () => {
   // NOTE: We're explicitly not using JSX here. This is intended to test
   // classic JS without JSX.
   it('identifies valid elements', () => {
-    var Component = React.createClass({
-      render: function() {
+    class Component extends React.Component {
+      render() {
         return React.createElement('div');
-      },
-    });
+      }
+    }
 
     expect(React.isValidElement(React.createElement('div')))
       .toEqual(true);
@@ -357,21 +320,6 @@ describe('ReactElement', () => {
     expect(React.isValidElement(JSON.parse(jsonElement))).toBe(true);
   });
 
-  it('allows the use of PropTypes validators in statics', () => {
-    // TODO: This test was added to cover a special case where we proxied
-    // methods. However, we don't do that any more so this test can probably
-    // be removed. Leaving it in classic as a safety precaution.
-    var Component = React.createClass({
-      render: () => null,
-      statics: {
-        specialType: React.PropTypes.shape({monkey: React.PropTypes.any}),
-      },
-    });
-
-    expect(typeof Component.specialType).toBe('function');
-    expect(typeof Component.specialType.isRequired).toBe('function');
-  });
-
   // NOTE: We're explicitly not using JSX here. This is intended to test
   // classic JS without JSX.
   it('is indistinguishable from a plain object', () => {
@@ -383,14 +331,12 @@ describe('ReactElement', () => {
   // NOTE: We're explicitly not using JSX here. This is intended to test
   // classic JS without JSX.
   it('should use default prop value when removing a prop', () => {
-    var Component = React.createClass({
-      getDefaultProps: function() {
-        return {fruit: 'persimmon'};
-      },
-      render: function() {
+    class Component extends React.Component {
+      render() {
         return React.createElement('span');
-      },
-    });
+      }
+    }
+    Component.defaultProps = {fruit: 'persimmon'};
 
     var container = document.createElement('div');
     var instance = ReactDOM.render(
@@ -406,14 +352,12 @@ describe('ReactElement', () => {
   // NOTE: We're explicitly not using JSX here. This is intended to test
   // classic JS without JSX.
   it('should normalize props with default values', () => {
-    var Component = React.createClass({
-      getDefaultProps: function() {
-        return {prop: 'testKey'};
-      },
-      render: function() {
+    class Component extends React.Component {
+      render() {
         return React.createElement('span', null, this.props.prop);
-      },
-    });
+      }
+    }
+    Component.defaultProps = {prop: 'testKey'};
 
     var instance = ReactTestUtils.renderIntoDocument(
       React.createElement(Component)
@@ -427,8 +371,8 @@ describe('ReactElement', () => {
   });
 
   it('throws when changing a prop (in dev) after element creation', () => {
-    var Outer = React.createClass({
-      render: function() {
+    class Outer extends React.Component {
+      render() {
         var el = <div className="moo" />;
 
         expect(function() {
@@ -437,17 +381,16 @@ describe('ReactElement', () => {
         expect(el.props.className).toBe('moo');
 
         return el;
-      },
-    });
+      }
+    }
     var outer = ReactTestUtils.renderIntoDocument(<Outer color="orange" />);
     expect(ReactDOM.findDOMNode(outer).className).toBe('moo');
   });
 
   it('throws when adding a prop (in dev) after element creation', () => {
     var container = document.createElement('div');
-    var Outer = React.createClass({
-      getDefaultProps: () => ({sound: 'meow'}),
-      render: function() {
+    class Outer extends React.Component {
+      render() {
         var el = <div>{this.props.sound}</div>;
 
         expect(function() {
@@ -457,8 +400,9 @@ describe('ReactElement', () => {
         expect(el.props.className).toBe(undefined);
 
         return el;
-      },
-    });
+      }
+    }
+    Outer.defaultProps = {sound: 'meow'};
     var outer = ReactDOM.render(<Outer />, container);
     expect(ReactDOM.findDOMNode(outer).textContent).toBe('meow');
     expect(ReactDOM.findDOMNode(outer).className).toBe('');
@@ -466,11 +410,11 @@ describe('ReactElement', () => {
 
   it('does not warn for NaN props', () => {
     spyOn(console, 'error');
-    var Test = React.createClass({
-      render: function() {
+    class Test extends React.Component {
+      render() {
         return <div />;
-      },
-    });
+      }
+    }
     var test = ReactTestUtils.renderIntoDocument(<Test value={+undefined} />);
     expect(test.props.value).toBeNaN();
     expect(console.error.calls.count()).toBe(0);
@@ -498,11 +442,11 @@ describe('ReactElement', () => {
 
     React = require('React');
 
-    var Component = React.createClass({
-      render: function() {
+    class Component extends React.Component {
+      render() {
         return React.createElement('div');
-      },
-    });
+      }
+    }
 
     expect(React.isValidElement(React.createElement('div')))
       .toEqual(true);
@@ -538,15 +482,15 @@ describe('comparing jsx vs .createFactory() vs .createElement()', () => {
   describe('when using jsx only', () => {
     var Parent, instance;
     beforeEach(() => {
-      Parent = React.createClass({
-        render: function() {
+      Parent = class extends React.Component {
+        render() {
           return (
             <div>
               <Child ref="child" foo="foo value">children value</Child>
             </div>
           );
-        },
-      });
+        }
+      };
       instance = ReactTestUtils.renderIntoDocument(<Parent/>);
     });
 
@@ -568,11 +512,11 @@ describe('comparing jsx vs .createFactory() vs .createElement()', () => {
     var factory, instance;
     beforeEach(() => {
       var childFactory = React.createFactory(Child);
-      var Parent = React.createClass({
-        render: function() {
+      class Parent extends React.Component {
+        render() {
           return React.DOM.div({}, childFactory({ ref: 'child', foo: 'foo value' }, 'children value'));
-        },
-      });
+        }
+      }
       factory = React.createFactory(Parent);
       instance = ReactTestUtils.renderIntoDocument(factory());
     });
@@ -594,11 +538,11 @@ describe('comparing jsx vs .createFactory() vs .createElement()', () => {
   describe('when using parent that uses .createElement()', () => {
     var factory, instance;
     beforeEach(() => {
-      var Parent = React.createClass({
-        render: function() {
+      class Parent extends React.Component {
+        render() {
           return React.DOM.div({}, React.createElement(Child, { ref: 'child', foo: 'foo value' }, 'children value'));
-        },
-      });
+        }
+      }
       factory = React.createFactory(Parent);
       instance = ReactTestUtils.renderIntoDocument(factory());
     });
