@@ -391,30 +391,23 @@ describe('ReactUpdates', () => {
       },
     };
 
-    var Box = React.createClass({
-      mixins: [UpdateLoggingMixin],
-
-      render: function() {
+    class Box extends React.Component {
+      render() {
         return <div ref="boxDiv">{this.props.children}</div>;
-      },
-    });
+      }
+    }
+    Object.assign(Box.prototype, UpdateLoggingMixin);
 
-    var Child = React.createClass({
-      mixins: [UpdateLoggingMixin],
-
-      render: function() {
+    class Child extends React.Component {
+      render() {
         return <span ref="span">child</span>;
-      },
-    });
+      }
+    }
+    Object.assign(Child.prototype, UpdateLoggingMixin);
 
-    var Switcher = React.createClass({
-      mixins: [UpdateLoggingMixin],
-
-      getInitialState: function() {
-        return {tabKey: 'hello'};
-      },
-
-      render: function() {
+    class Switcher extends React.Component {
+      state = {tabKey: 'hello'};
+      render() {
         var child = this.props.children;
 
         return (
@@ -428,20 +421,20 @@ describe('ReactUpdates', () => {
             </div>
           </Box>
         );
-      },
-    });
+      }
+    }
+    Object.assign(Switcher.prototype, UpdateLoggingMixin);
 
-    var App = React.createClass({
-      mixins: [UpdateLoggingMixin],
-
-      render: function() {
+    class App extends React.Component {
+      render() {
         return (
           <Switcher ref="switcher">
             <Child key="hello" ref="child" />
           </Switcher>
         );
-      },
-    });
+      }
+    }
+    Object.assign(App.prototype, UpdateLoggingMixin);
 
     var root = <App />;
     root = ReactTestUtils.renderIntoDocument(root);
@@ -984,35 +977,6 @@ describe('ReactUpdates', () => {
     );
     expect(() => component.setState({}, new Foo())).toThrowError(
       'setState(...): Expected the last optional `callback` argument ' +
-      'to be a function. Instead received: Foo (keys: a, b).'
-    );
-  });
-
-  it('throws in replaceState if the update callback is not a function', () => {
-    function Foo() {
-      this.a = 1;
-      this.b = 2;
-    }
-    var A = React.createClass({
-      getInitialState: function() {
-        return {};
-      },
-      render: function() {
-        return <div />;
-      },
-    });
-    var component = ReactTestUtils.renderIntoDocument(<A />);
-
-    expect(() => component.replaceState({}, 'no')).toThrowError(
-      'replaceState(...): Expected the last optional `callback` argument ' +
-      'to be a function. Instead received: string.'
-    );
-    expect(() => component.replaceState({}, {})).toThrowError(
-      'replaceState(...): Expected the last optional `callback` argument ' +
-      'to be a function. Instead received: Object.'
-    );
-    expect(() => component.replaceState({}, new Foo())).toThrowError(
-      'replaceState(...): Expected the last optional `callback` argument ' +
       'to be a function. Instead received: Foo (keys: a, b).'
     );
   });
