@@ -14,12 +14,12 @@
 var ReactWithAddons = require('ReactWithAddons');
 
 // `version` will be added here by the React module.
-var ReactWithAddonsUMDEntry = Object.assign({
+var ReactWithAddonsUMDEntry = Object.assign(ReactWithAddons, {
   __SECRET_INJECTED_REACT_DOM_DO_NOT_USE_OR_YOU_WILL_BE_FIRED: null, // Will be injected by ReactDOM UMD build.
   __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED: {
     ReactCurrentOwner: require('ReactCurrentOwner'),
   },
-}, ReactWithAddons);
+});
 
 if (__DEV__) {
   Object.assign(
@@ -27,6 +27,7 @@ if (__DEV__) {
     {
       // ReactComponentTreeHook should not be included in production.
       ReactComponentTreeHook: require('ReactComponentTreeHook'),
+      getNextDebugID: require('getNextDebugID'),
     }
   );
 }
